changes switch in main and added filter to MainCarousel
