+++ src/Components/DisplayRecipes/DisplayRecipes.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function DisplayRecipes() {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default DisplayRecipes
